+++ domain/src/main/java/com/agileburo/anytype/domain/block/model/Command.kt
@@ -88,6 +88,20 @@ sealed class Command {
         val prototype: Block.Prototype
     )
 
+    /**
+     * Command for creating a new document / page
+     * @property context id of the context of the block (i.e. page, dashboard or something else)
+     * @property target id of the block associated with the block we need to create
+     * @property position position of the block that we need to create in relation with the target block
+     * @property prototype a prototype of the block we would like to create
+     */
+    class CreateDocument(
+        val context: Id,
+        val target: Id,
+        val position: Position,
+        val prototype: Block.Prototype.Page
+    )
+
     class Dnd(
         val contextId: Id,
         val targetId: Id,