+++ middleware/src/test/java/com/agileburo/anytype/MiddlewareTest.kt
@@ -1,17 +1,22 @@
 package com.agileburo.anytype
 
 import anytype.Commands.Rpc.Account
+import anytype.Commands.Rpc.Block
+import anytype.model.Models
+import com.agileburo.anytype.common.MockDataFactory
+import com.agileburo.anytype.data.auth.model.BlockEntity
+import com.agileburo.anytype.data.auth.model.CommandEntity
+import com.agileburo.anytype.data.auth.model.PositionEntity
 import com.agileburo.anytype.middleware.interactor.Middleware
 import com.agileburo.anytype.middleware.interactor.MiddlewareFactory
 import com.agileburo.anytype.middleware.interactor.MiddlewareMapper
 import com.agileburo.anytype.middleware.service.MiddlewareService
-import com.nhaarman.mockitokotlin2.times
-import com.nhaarman.mockitokotlin2.verify
-import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
+import com.nhaarman.mockitokotlin2.*
 import org.junit.Before
 import org.junit.Test
 import org.mockito.Mock
 import org.mockito.MockitoAnnotations
+import kotlin.test.assertEquals
 
 class MiddlewareTest {
 
@@ -37,5 +42,57 @@ class MiddlewareTest {
 
         verify(service, times(1)).accountStop(request)
         verifyNoMoreInteractions(service)
+    }
+
+    @Test
+    fun `should create request to create new document and return pair of ids`() {
+
+        val command = CommandEntity.CreateDocument(
+            context = MockDataFactory.randomUuid(),
+            target = MockDataFactory.randomUuid(),
+            prototype = BlockEntity.Prototype.Page(
+                style = BlockEntity.Content.Page.Style.EMPTY
+            ),
+            position = PositionEntity.INNER
+        )
+
+        val response = Block.CreatePage.Response
+            .newBuilder()
+            .setBlockId(MockDataFactory.randomUuid())
+            .setTargetId(MockDataFactory.randomUuid())
+            .build()
+
+        val request = Block.CreatePage.Request
+            .newBuilder()
+            .setContextId(command.context)
+            .setTargetId(command.target)
+            .setPosition(Models.Block.Position.Inner)
+            .setBlock(
+                Models.Block.newBuilder()
+                    .setPage(
+                        Models.Block.Content.Page
+                            .newBuilder()
+                            .setStyle(Models.Block.Content.Page.Style.Empty)
+                    )
+            )
+            .build()
+
+        service.stub {
+            on { blockCreatePage(any()) } doReturn response
+        }
+
+        val (block, target) = middleware.createDocument(command)
+
+        verify(service, times(1)).blockCreatePage(request)
+
+        assertEquals(
+            expected = response.blockId,
+            actual = block
+        )
+
+        assertEquals(
+            expected = response.targetId,
+            actual = target
+        )
     }
 }