+++ presentation/src/test/java/com/agileburo/anytype/presentation/page/DocumentExternalEventReducerTest.kt
@@ -2,6 +2,7 @@ package com.agileburo.anytype.presentation.page
 
 import MockDataFactory
 import com.agileburo.anytype.domain.block.model.Block
+import com.agileburo.anytype.domain.block.model.Block.Fields.Companion.NAME_KEY
 import com.agileburo.anytype.domain.event.model.Event
 import com.agileburo.anytype.domain.ext.content
 import kotlinx.coroutines.runBlocking
@@ -207,6 +208,78 @@ class DocumentExternalEventReducerTest {
                     description = null,
                     favicon = null,
                     image = null
+                )
+            )
+
+            assertEquals(expected = expected, actual = result)
+        }
+    }
+
+    @Test
+    fun `should update link fields`() {
+
+        val title = Block(
+            id = MockDataFactory.randomUuid(),
+            fields = Block.Fields.empty(),
+            children = emptyList(),
+            content = Block.Content.Text(
+                text = MockDataFactory.randomString(),
+                marks = emptyList(),
+                style = Block.Content.Text.Style.TITLE
+            )
+        )
+
+        val name = MockDataFactory.randomString()
+
+        val link = Block(
+            id = MockDataFactory.randomUuid(),
+            fields = Block.Fields.empty(),
+            children = emptyList(),
+            content = Block.Content.Link(
+                fields = Block.Fields(
+                    map = mapOf(NAME_KEY to name)
+                ),
+                target = MockDataFactory.randomUuid(),
+                type = Block.Content.Link.Type.PAGE
+            )
+        )
+
+        val page = Block(
+            id = MockDataFactory.randomUuid(),
+            fields = Block.Fields.empty(),
+            children = listOf(title.id, link.id),
+            content = Block.Content.Page(
+                style = Block.Content.Page.Style.SET
+            )
+        )
+
+        val state = listOf(page, title, link)
+
+        // TESTING
+
+        val updated = Block.Fields(
+            map = mapOf(NAME_KEY to MockDataFactory.randomString())
+        )
+
+        runBlocking {
+
+            val expected = listOf(
+                page,
+                title,
+                link.copy(
+                    content = link.content<Block.Content.Link>().copy(
+                        fields = updated
+                    )
+                )
+            )
+
+            val result = reducer.reduce(
+                state = state,
+                event = Event.Command.LinkGranularChange(
+                    context = page.id,
+                    id = link.id,
+                    target = link.content<Block.Content.Link>().target,
+                    fields = updated
                 )
             )